+++ tests/examples/example-bloomfilter.c
@@ -4,8 +4,8 @@
 #include <twiddle/bloomfilter/bloomfilter.h>
 
 int main(int argc, char *argv[]) {
-  const uint32_t nbits = 1024;
-  const uint32_t k = 7;
+  const uint64_t nbits = 1024;
+  const uint16_t k = 7;
   struct tw_bloomfilter *bf = tw_bloomfilter_new(nbits, k);
   assert(bf);
